day_1/pyproject.toml
[project]
name = "04-producton-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.12.*"
dependencies = [
    "jupyter>=1.1.1",
    "langchain-community>=0.3.26",
    "langchain-core>=0.3.67",
    "langchain-openai>=0.3.27",
    "langchain-qdrant>=0.2.0",
    "langgraph>=0.5.0",
    "langsmith>=0.4.4",
    "marimo>=0.17.0",
    "pandas>=2.3.0",
    "pymupdf>=1.26.1",
]
